Fix claimRewards, MyRewards fix date, my metrics
- Connected the claim rewards context menu to hit the api
- Made updates so that the moment functions arn't used for my rewards
- Changed the way conversion rates are handled for my metrics
